services/aesthete-engine/requirements.txt
# FastAPI and web server
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# ML and embeddings
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
onnxruntime==1.17.0
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Vector similarity and search
pgvector==0.2.4
psycopg2-binary==2.9.9
opensearch-py==2.4.2

# Database and ORM
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# MLflow for model registry
mlflow==2.10.2
boto3==1.34.34  # for S3/OCI Object Storage artifacts

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-json-logger==2.0.7
python-multipart==0.0.9
pydantic-core==2.16.2
python-dotenv==1.0.1
pyyaml==6.0.1
